Move Bm25ForLabelsExecutor into the cpp file's anonymous namespace
The executor is a pure implementation detail.  Relocating the class into
the anonymous namespace in bm25_for_labels_feature.cpp shrinks the header
down to just the Bm25ForLabelsBlueprint interface, improving encapsulation
and cutting the header's include/compile footprint. No behavioural change.
